Lookup from postcode to street name by LA
This file contains provisional allocation of postcodes to individual named streets in Great Britain. Named streets can be disambiguated with the nametoid column which contains an OSGB toid for the entire length of the street. An ONS local authority code is also provided for filtering streets of interest.

Allocation is based on OS Open Road (for road geometries and nametoids), the OS Open Postcode dataset and ONS UPRN lookups. No OpenStreetMap data was used. Allocation is mainly based on proximity of a postcode centroid to various road geometries.